+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/web/controller/CertifiedProductControllerTest.java
@@ -1140,6 +1140,9 @@ public class CertifiedProductControllerTest {
 				}
 			}
 			assertTrue(e.getContact() != null);
+		} catch (IOException e) {
+			// TODO Auto-generated catch block
+			e.printStackTrace();
 		}
 		assertTrue(hasError);
 	}